server/requirements.txt
asgiref==3.4.1
boto3==1.18.11
botocore==1.21.51
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.5
djangorestframework==3.12.4
django-heroku==0.3.1
django-storages==1.11.1
django-cors-headers==3.7.0
dnspython==2.1.0
gunicorn==20.1.0
itypes==1.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyjwt==2.1.0
pymongo==3.12.0
pytest==6.2.5
python-dotenv==0.19.0
pytz==2021.1
pyyaml==5.4
requests==2.26.0
s3transfer==0.5.0
sqlparse==0.4.2
uritemplate==4.1.1